+++ changes.md
@@ -2,6 +2,7 @@
 * Moved MU plugins to be loaded from ./application
 * Added autoload.php MU plugin for loading vendor libs
 * Re-installed wp-cluster, wp-veneer and wp-network as valid plugins, removing ./vendor/modules completely.
+* Removed automatic loading of wp-vertical-edm and wp-elastic, should be enabled as a plugin.
 
 #### 2.1.3
 * Added changes.md